Remove long double

+++ include/xieite/geometry/Line.hpp
@@ -11,12 +11,12 @@ namespace xieite::geometry {
 		constexpr Line(const xieite::geometry::Point start, const xieite::geometry::Point end) noexcept
 		: xieite::geometry::LineLike(start, end) {}
 
-		constexpr Line(const xieite::geometry::Point start, const long double angle) noexcept
+		constexpr Line(const xieite::geometry::Point start, const double angle) noexcept
 		: xieite::geometry::LineLike(start, xieite::geometry::Point(std::cos(angle), std::sin(angle))) {}
 
 		[[nodiscard]]
 		constexpr bool operator==(const xieite::geometry::Line& other) const noexcept {
-			const long double slope = xieite::geometry::getSlope(*this);
+			const double slope = xieite::geometry::getSlope(*this);
 			return (std::isinf(slope) ? xieite::math::approximatelyEqual(other.start.x, start.x) : xieite::math::approximatelyEqual(other.start.x * slope - start.x * slope + start.y, other.start.y)) && xieite::math::approximatelyEqual(slope, xieite::geometry::getSlope(other));
 		}
 	};
